pyproject.toml
[tool.poetry]
name = "spam_finder"
version = "0.1.0"
description = ""
authors = ["Rubab Uddin <uddin.rubab@gmail.com>", "Nathan Sass <nsass711@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
python-dotenv = "1.0.1"
langchain = ">=0.1.0,<0.2.0"
langchain-openai = ">=0.0.5,<0.1.8"
langchain-community = ">0.0.30,<0.2.4"
crewai = { version = "^0.30.11", python = ">=3.11,<=3.13" }
crewai-tools = { version = "^0.2.6" }
chromadb = ">=0.4.22,<0.5.0"
youtube-transcript-api = "^0.6.2"
google = "^3.0.0"
google-api-python-client = "^2.130.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
nltk = "^3.8.1"
uvicorn = "0.30.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"